[Example] Change: Extract settings
