+++ example/app/example/ExampleOverlayImage.tsx
@@ -2,8 +2,8 @@ import * as React from "react"
 import { Image } from "react-native"
 import styles from "../App.styles"
 import PhotoManipulator from "../../../src/PhotoManipulator"
-import { IMAGE, OVERLAY } from "./index"
 import { noop } from "../utils"
+import { IMAGE, OVERLAY } from "./settings"
 
 export default React.memo(function ExampleOverlayImage() {
   const [image, setImage] = React.useState<string|null>(null)
