+++ example/app/example/index.tsx
@@ -3,6 +3,7 @@ import { Image } from "react-native"
 import styles from "../App.styles"
 import ExampleOverlayImage from "./ExampleOverlayImage"
 import ExamplePrintText from "./ExamplePrintText"
+import { IMAGE, OVERLAY } from "./settings"
 
 export interface Example {
   id: string
@@ -10,9 +11,6 @@ export interface Example {
   description: string
   render: () => React.ReactNode
 }
-
-export const IMAGE = "https://images.pexels.com/photos/67636/rose-blue-flower-rose-blooms-67636.jpeg?cs=srgb&dl=beauty-bloom-blue-67636.jpg&fm=jpg?dl&fit=crop&crop=entropy&w=800&h="
-export const OVERLAY = "https://upload.wikimedia.org/wikipedia/commons/thumb/a/a7/React-icon.svg/200px-React-icon.svg.png"
 
 const EXAMPLES: Example[] = [
   {
