+++ example/app/example/settings.ts
@@ -1,0 +1,2 @@
+export const IMAGE = "https://images.pexels.com/photos/67636/rose-blue-flower-rose-blooms-67636.jpeg?cs=srgb&dl=beauty-bloom-blue-67636.jpg&fm=jpg?dl&fit=crop&crop=entropy&w=800&h="
+export const OVERLAY = "https://upload.wikimedia.org/wikipedia/commons/thumb/a/a7/React-icon.svg/200px-React-icon.svg.png"
